chore(alert_muting_rule): Organizing structures file.
